requirements.txt
discord.py~=2.1
websockets~=10.4
psutil~=5.9.4
wavelink==0.9.10
asyncurban~=0.3.3
aiohttp==3.7.4
asyncpg~=0.27.0
asyncpraw~=7.6.0
beautifulsoup4~=4.11.1
setproctitle~=1.3.2
requests~=2.28.1
topggpy~=1.4.0
aioredis~=2.0.1
multiprocessing-logging~=0.3.3